move Locate (old attendee_contact) to whereabouts app
