+++ attendees/persons/models/__init__.py
@@ -10,4 +10,3 @@ from .relationship import Relationship
 from .registration import Registration
 from .attending import Attending
 from .attending_meet import AttendingMeet
-from .locate import Locate
