+++ attendees/whereabouts/models/__init__.py
@@ -1,6 +1,7 @@
 from .organization import Organization
 from .division import Division
 from .place import Place
+from .locate import Locate
 from .campus import Campus
 from .property import Property
 from .suite import Suite
